requirements.txt
requests
flask
flask-bootstrap
flask-mongoengine
flask-script
flask-security
flask-login
jinja2
bcrypt
